add the obfuscation-feature
The shortcode allows you to mask sensitive information such as email addresses, phone numbers, or text, for example, an address.

Usable Parameters

When using the shortcode, the following parameters must be specified:
mail: The email address to be obfuscated.
tel: The phone number to be obfuscated.
text: The text, for example, an address, to be obfuscated.
title: The title to be displayed as text on the button.
class: (optional) CSS class to customize the button.

Example Usage

To use the shortcode, it should be entered as follows:

[obfuscation mail="example@mail.com" title="Show Mail" class="myClass"]

Important Notes

Required Parameters: At least one of the parameters mail, tel, or text must be specified. The title is always required.
Optional Customization: The class parameter is optional and allows you to customize the button using CSS.